[rb] remove remaining Edge workarounds for remote server
